backend/app/requirements.txt
fastapi==0.99.1
uvicorn==0.22.0
pydantic==1.10.9
sqlalchemy==1.4.49
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
